Minor changes
HTML Changes
In this push I reorganised the social bar at the bottom into something that I thought fitted better.
Also added a section for two boxes to be added when needed inside the content box (will fully be implemented when the pages need it)

CSS
I also fixed the positioning so that doesn't look weird depending on your browser zoom.
and I also redid the ordering in the CSS so that the main divs are at the top of the CSS (made it a lot easier for the positioning)